(refactor) changed filename to fit React naming conventions
